+++ ocd/packages/export/src/terraform/OcdTerraformResource.ts
@@ -19,22 +19,27 @@ export class OcdTerraformResource {
     generateTextAttribute = (name: string, value: string | undefined, required: boolean) => {
         if (this.isVariable(value)) return `${name} = ${this.formatVariable(value)}`
         else if (required) return `${name} = "${value}"`
-        else if (value && value !== '') return `${name} = "${value}"`
+        else if (value && value.trim() !== '') return `${name} = "${value}"`
         else return `# ${name} = "${value}"`
     }
     generateBooleanAttribute = (name: string, value: string | boolean | undefined, required: boolean) => {
-        console.debug('OcdTerraformResource: generateBooleanAttribute:', value, typeof value)
         if (typeof value === 'string' && this.isVariable(value)) return `${name} = ${this.formatVariable(value)}`
         else if (required) return `${name} = ${value}`
         else if (typeof value === 'boolean') return `${name} = ${value}`
         else return `# ${name} = ${value}`
     }
     generateNumberAttribute = (name: string, value: string | number | undefined, required: boolean) => {
-        console.debug('OcdTerraformResource: generateNumberAttribute:', value, typeof value)
         if (typeof value === 'string' && this.isVariable(value)) return `${name} = ${this.formatVariable(value)}`
         else if (required) return `${name} = ${value}`
         else if (value && typeof value === 'number') return `${name} = ${value}`
         else return `# ${name} = ${value}`
+    }
+    generateReferenceListAttribute = (name: string, value: string | string[] | undefined, required: boolean) => {
+        console.debug('OcdTerraformResource: generateReferenceListAttribute:', value, typeof value)
+        if (!Array.isArray(value) && this.isVariable(value)) return `${name} = ${this.formatVariable(value as string)}`
+        else if (required && Array.isArray(value)) return `${name} = [${value.map((v: string) => `local.${this.idTFResourceMap[v]}_id`)}]`
+        else if (Array.isArray(value) && value.length > 0) return `${name} = [${value.map((v: string) => `local.${this.idTFResourceMap[v]}_id`)}]`
+        else return `# ${name} = "${value}"`
     }
 
     // if (attribute.type === 'string')      return `${name} = \\\`\${this.isVariable(resource.${this.toCamelCase(name)}) ? this.formatVariable(resource.${this.toCamelCase(name)}) : "resource.${this.toCamelCase(name)}"}\\\``